services/config_manager_service/requirements.txt
# Core Dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.6.1
structlog>=24.1.0
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.30b1
pyyaml>=6.0
aiofiles>=0.8.0
watchfiles>=0.18.0
jsonschema>=4.17.0

# Database
sqlalchemy[asyncio]>=2.0.27
alembic>=1.13.1
asyncpg>=0.29.0 